// lib/powerfox-sdk/models/MyDeviceModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kunden API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Divisions } from './Divisions';
import {
    DivisionsFromJSON,
    DivisionsFromJSONTyped,
    DivisionsToJSON,
    DivisionsToJSONTyped,
} from './Divisions';

/**
 * 
 * @export
 * @interface MyDeviceModel
 */
export interface MyDeviceModel {
    /**
     * 
     * @type {string}
     * @memberof MyDeviceModel
     */
    deviceId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MyDeviceModel
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof MyDeviceModel
     */
    accountAssociatedSince?: number;
    /**
     * 
     * @type {boolean}
     * @memberof MyDeviceModel
     */
    mainDevice?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof MyDeviceModel
     */
    prosumer?: boolean;
    /**
     * 
     * @type {Divisions}
     * @memberof MyDeviceModel
     */
    division?: Divisions;
}



/**
 * Check if a given object implements the MyDeviceModel interface.
 */
export function instanceOfMyDeviceModel(value: object): value is MyDeviceModel {
    return true;
}

export function MyDeviceModelFromJSON(json: any): MyDeviceModel {
    return MyDeviceModelFromJSONTyped(json, false);
}

export function MyDeviceModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): MyDeviceModel {
    if (json == null) {
        return json;
    }
    return {
        
        'deviceId': json['DeviceId'] == null ? undefined : json['DeviceId'],
        'name': json['Name'] == null ? undefined : json['Name'],
        'accountAssociatedSince': json['AccountAssociatedSince'] == null ? undefined : json['AccountAssociatedSince'],
        'mainDevice': json['MainDevice'] == null ? undefined : json['MainDevice'],
        'prosumer': json['Prosumer'] == null ? undefined : json['Prosumer'],
        'division': json['Division'] == null ? undefined : DivisionsFromJSON(json['Division']),
    };
}

export function MyDeviceModelToJSON(json: any): MyDeviceModel {
    return MyDeviceModelToJSONTyped(json, false);
}

export function MyDeviceModelToJSONTyped(value?: MyDeviceModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'DeviceId': value['deviceId'],
        'Name': value['name'],
        'AccountAssociatedSince': value['accountAssociatedSince'],
        'MainDevice': value['mainDevice'],
        'Prosumer': value['prosumer'],
        'Division': DivisionsToJSON(value['division']),
    };
}