// lib/powerfox-sdk/models/MyCurrentDataModel.ts
export interface MyCurrentDataModel {
    /**
     * 
     * @type {string}
     * @memberof MyCurrentDataModel
     */
    deviceId?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof MyCurrentDataModel
     */
    outdated?: boolean;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    watt?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    kiloWattHour?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    deltaKiloWattHour?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    cubicMeterCold?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    cubicMeterWarm?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    cubicMeter?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    deltaCubicMeter?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    timestamp?: number;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    aPlus?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    aPlusHT?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    aPlusNT?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    aMinus?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    l1?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    l2?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MyCurrentDataModel
     */
    l3?: number | null;
}