// lib/powerfox-sdk/models/OperatingReportModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kunden API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OperatingReportValueModel } from './OperatingReportValueModel';
import {
    OperatingReportValueModelFromJSON,
    OperatingReportValueModelFromJSONTyped,
    OperatingReportValueModelToJSON,
    OperatingReportValueModelToJSONTyped,
} from './OperatingReportValueModel';

/**
 * 
 * @export
 * @interface OperatingReportModel
 */
export interface OperatingReportModel {
    /**
     * 
     * @type {string}
     * @memberof OperatingReportModel
     */
    deviceId?: string | null;
    /**
     * 
     * @type {Array<OperatingReportValueModel>}
     * @memberof OperatingReportModel
     */
    values?: Array<OperatingReportValueModel> | null;
    /**
     * 
     * @type {Array<OperatingReportValueModel>}
     * @memberof OperatingReportModel
     */
    valuesPlus?: Array<OperatingReportValueModel> | null;
    /**
     * 
     * @type {Array<OperatingReportValueModel>}
     * @memberof OperatingReportModel
     */
    valuesMinus?: Array<OperatingReportValueModel> | null;
    /**
     * 
     * @type {number}
     * @memberof OperatingReportModel
     */
    max?: number;
    /**
     * 
     * @type {number}
     * @memberof OperatingReportModel
     */
    min?: number;
    /**
     * 
     * @type {number}
     * @memberof OperatingReportModel
     */
    avg?: number;
}

/**
 * Check if a given object implements the OperatingReportModel interface.
 */
export function instanceOfOperatingReportModel(value: object): value is OperatingReportModel {
    return true;
}

export function OperatingReportModelFromJSON(json: any): OperatingReportModel {
    return OperatingReportModelFromJSONTyped(json, false);
}

export function OperatingReportModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): OperatingReportModel {
    if (json == null) {
        return json;
    }
    return {
        
        'deviceId': json['DeviceId'] == null ? undefined : json['DeviceId'],
        'values': json['Values'] == null ? undefined : ((json['Values'] as Array<any>).map(OperatingReportValueModelFromJSON)),
        'valuesPlus': json['ValuesPlus'] == null ? undefined : ((json['ValuesPlus'] as Array<any>).map(OperatingReportValueModelFromJSON)),
        'valuesMinus': json['ValuesMinus'] == null ? undefined : ((json['ValuesMinus'] as Array<any>).map(OperatingReportValueModelFromJSON)),
        'max': json['Max'] == null ? undefined : json['Max'],
        'min': json['Min'] == null ? undefined : json['Min'],
        'avg': json['Avg'] == null ? undefined : json['Avg'],
    };
}

export function OperatingReportModelToJSON(json: any): OperatingReportModel {
    return OperatingReportModelToJSONTyped(json, false);
}

export function OperatingReportModelToJSONTyped(value?: OperatingReportModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'DeviceId': value['deviceId'],
        'Values': value['values'] == null ? undefined : ((value['values'] as Array<any>).map(OperatingReportValueModelToJSON)),
        'ValuesPlus': value['valuesPlus'] == null ? undefined : ((value['valuesPlus'] as Array<any>).map(OperatingReportValueModelToJSON)),
        'ValuesMinus': value['valuesMinus'] == null ? undefined : ((value['valuesMinus'] as Array<any>).map(OperatingReportValueModelToJSON)),
        'Max': value['max'],
        'Min': value['min'],
        'Avg': value['avg'],
    };
}