// lib/powerfox-sdk/models/ReportModel.ts
export interface ReportModel {
    /**
     * 
     * @type {string}
     * @memberof ReportModel
     */
    deviceId?: string | null;
    /**
     * 
     * @type {AppReportSummaryPower}
     * @memberof ReportModel
     */
    consumption?: AppReportSummaryPower;
    /**
     * 
     * @type {AppReportSummaryPower}
     * @memberof ReportModel
     */
    ownConsumption?: AppReportSummaryPower;
    /**
     * 
     * @type {AppReportSummaryPower}
     * @memberof ReportModel
     */
    feedIn?: AppReportSummaryPower;
    /**
     * 
     * @type {AppReportSummaryPower}
     * @memberof ReportModel
     */
    generation?: AppReportSummaryPower;
    /**
     * 
     * @type {AppReportSummaryHeat}
     * @memberof ReportModel
     */
    heat?: AppReportSummaryHeat;
    /**
     * 
     * @type {AppReportSummaryGas}
     * @memberof ReportModel
     */
    gas?: AppReportSummaryGas;
    /**
     * 
     * @type {AppReportSummaryWater}
     * @memberof ReportModel
     */
    water?: AppReportSummaryWater;
}